+++ collections/indexed_map.go
@@ -1,15 +1,12 @@
 package collections
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/NibiruChain/nibiru/collections/keys"
 )
 
 // IndexersProvider is implemented by structs containing
 // a series of Indexer instances.
-type IndexersProvider[PK keys.Key, V any] interface {
+type IndexersProvider[PK any, V any] interface {
 	// IndexerList provides the list of Indexer contained
 	// in the struct.
 	IndexerList() []Indexer[PK, V]
@@ -19,7 +16,7 @@ type IndexersProvider[PK keys.Key, V any] interface {
 // and a primary key PK, creates a relationship
 // between one or multiple fields of the object V
 // with the primary key PK.
-type Indexer[PK keys.Key, V any] interface {
+type Indexer[PK any, V any] interface {
 	// Insert is called when the IndexedMap is inserting
 	// an object into its state, so the Indexer here
 	// creates the relationship between primary key
@@ -32,12 +29,13 @@ type Indexer[PK keys.Key, V any] interface {
 }
 
 // NewIndexedMap instantiates a new IndexedMap instance.
-func NewIndexedMap[PK keys.Key, V any, PV interface {
-	*V
-	Object
-}, I IndexersProvider[PK, V]](cdc codec.BinaryCodec, storeKey sdk.StoreKey, namespace uint8, indexers I) IndexedMap[PK, V, PV, I] {
-	m := NewMap[PK, V, PV](cdc, storeKey, namespace)
-	return IndexedMap[PK, V, PV, I]{
+func NewIndexedMap[PK any, V any, I IndexersProvider[PK, V]](
+	storeKey sdk.StoreKey, namespace Namespace,
+	primaryKeyEncoder KeyEncoder[PK],
+	valueEncoder ValueEncoder[V],
+	indexers I) IndexedMap[PK, V, I] {
+	m := NewMap[PK, V](storeKey, namespace, primaryKeyEncoder, valueEncoder)
+	return IndexedMap[PK, V, I]{
 		m:       m,
 		Indexes: indexers,
 	}
@@ -45,29 +43,26 @@ func NewIndexedMap[PK keys.Key, V any, PV interface {
 
 // IndexedMap defines a map which is indexed using the IndexersProvider
 // PK defines the primary key of the object V.
-type IndexedMap[PK keys.Key, V any, PV interface {
-	*V
-	Object
-}, I IndexersProvider[PK, V]] struct {
-	m       Map[PK, V, PV] // maintains PrimaryKey (PK) -> Object (V) bytes
-	Indexes I              // struct that groups together Indexer instances, implements IndexersProvider
+type IndexedMap[PK any, V any, I IndexersProvider[PK, V]] struct {
+	m       Map[PK, V] // maintains PrimaryKey (PK) -> Object (V) bytes
+	Indexes I          // struct that groups together Indexer instances, implements IndexersProvider
 }
 
 // Get returns the object V given its primary key PK.
-func (i IndexedMap[PK, V, PV, I]) Get(ctx sdk.Context, key PK) (V, error) {
+func (i IndexedMap[PK, V, I]) Get(ctx sdk.Context, key PK) (V, error) {
 	return i.m.Get(ctx, key)
 }
 
 // GetOr returns the object V given its primary key PK, or if the operation fails
 // returns the provided default.
-func (i IndexedMap[PK, V, PV, I]) GetOr(ctx sdk.Context, key PK, def V) V {
+func (i IndexedMap[PK, V, I]) GetOr(ctx sdk.Context, key PK, def V) V {
 	return i.m.GetOr(ctx, key, def)
 }
 
 // Insert inserts the object v into the Map using the primary key, then
 // iterates over every registered Indexer and instructs them to create
 // the relationship between the primary key PK and the object v.
-func (i IndexedMap[PK, V, PV, I]) Insert(ctx sdk.Context, key PK, v V) {
+func (i IndexedMap[PK, V, I]) Insert(ctx sdk.Context, key PK, v V) {
 	// before inserting we need to assert if another instance of this
 	// primary key exist in order to remove old relationships from indexes.
 	old, err := i.m.Get(ctx, key)
@@ -82,7 +77,7 @@ func (i IndexedMap[PK, V, PV, I]) Insert(ctx sdk.Context, key PK, v V) {
 // Delete fetches the object from the Map removes it from the Map
 // then instructs every Indexer to remove the relationships between
 // the object and the associated primary keys.
-func (i IndexedMap[PK, V, PV, I]) Delete(ctx sdk.Context, key PK) error {
+func (i IndexedMap[PK, V, I]) Delete(ctx sdk.Context, key PK) error {
 	// we prefetch the object
 	v, err := i.m.Get(ctx, key)
 	if err != nil {
@@ -99,17 +94,17 @@ func (i IndexedMap[PK, V, PV, I]) Delete(ctx sdk.Context, key PK) error {
 
 // Iterate iterates over the underlying store containing the concrete objects.
 // The range provided filters over the primary keys.
-func (i IndexedMap[PK, V, PV, I]) Iterate(ctx sdk.Context, rng keys.Range[PK]) MapIterator[PK, V, PV] {
+func (i IndexedMap[PK, V, I]) Iterate(ctx sdk.Context, rng Range[PK]) Iterator[PK, V] {
 	return i.m.Iterate(ctx, rng)
 }
 
-func (i IndexedMap[PK, V, PV, I]) index(ctx sdk.Context, key PK, v V) {
+func (i IndexedMap[PK, V, I]) index(ctx sdk.Context, key PK, v V) {
 	for _, indexer := range i.Indexes.IndexerList() {
 		indexer.Insert(ctx, key, v)
 	}
 }
 
-func (i IndexedMap[PK, V, PV, I]) unindex(ctx sdk.Context, key PK, v V) {
+func (i IndexedMap[PK, V, I]) unindex(ctx sdk.Context, key PK, v V) {
 	for _, indexer := range i.Indexes.IndexerList() {
 		indexer.Delete(ctx, key, v)
 	}