+++ collections/indexed_map_test.go
@@ -1,42 +1,37 @@
 package collections
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/NibiruChain/nibiru/collections/keys"
 )
 
 type person struct {
-	ID   keys.Uint64Key
-	City keys.StringKey
-}
-
-func (p person) Marshal() ([]byte, error) {
-	return json.Marshal(p)
-}
-
-func (p *person) Unmarshal(b []byte) error {
-	return json.Unmarshal(b, &p)
+	ID   uint64
+	City string
 }
 
 type indexes struct {
-	City MultiIndex[keys.StringKey, keys.Uint64Key, person]
+	City MultiIndex[string, uint64, person]
 }
 
-func (i indexes) IndexerList() []Indexer[keys.Uint64Key, person] {
-	return []Indexer[keys.Uint64Key, person]{i.City}
+func (i indexes) IndexerList() []Indexer[uint64, person] {
+	return []Indexer[uint64, person]{i.City}
 }
 
 func TestIndexedMap(t *testing.T) {
-	sk, ctx, cdc := deps()
-	m := NewIndexedMap[keys.Uint64Key, person, *person, indexes](cdc, sk, 0, indexes{
-		City: NewMultiIndex[keys.StringKey, keys.Uint64Key, person](cdc, sk, 1, func(v person) keys.StringKey {
-			return v.City
-		}),
-	})
+	sk, ctx, _ := deps()
+	m := NewIndexedMap[uint64, person, indexes](
+		sk, 0,
+		Uint64KeyEncoder, jsonValue[person]{},
+		indexes{
+			City: NewMultiIndex[string, uint64, person](sk, 1,
+				StringKeyEncoder, Uint64KeyEncoder,
+				func(v person) string {
+					return v.City
+				}),
+		},
+	)
 
 	m.Insert(ctx, 0, person{ID: 0, City: "milan"})
 	m.Insert(ctx, 1, person{ID: 1, City: "new york"})
@@ -44,13 +39,13 @@ func TestIndexedMap(t *testing.T) {
 
 	// correct insertion
 	res := m.Indexes.City.ExactMatch(ctx, "milan").PrimaryKeys()
-	require.Equal(t, []keys.Uint64Key{0, 2}, res)
+	require.Equal(t, []uint64{0, 2}, res)
 
 	// once deleted, it's removed from indexes
 	err := m.Delete(ctx, 0)
 	require.NoError(t, err)
 	res = m.Indexes.City.ExactMatch(ctx, "milan").PrimaryKeys()
-	require.Equal(t, []keys.Uint64Key{2}, res)
+	require.Equal(t, []uint64{2}, res)
 
 	// insertion on an already existing primary key
 	// clears the old indexes, hence PK 2 => city "milan"
@@ -58,7 +53,7 @@ func TestIndexedMap(t *testing.T) {
 	m.Insert(ctx, 2, person{ID: 2, City: "new york"})
 	require.Empty(t, m.Indexes.City.ExactMatch(ctx, "milan").PrimaryKeys())
 	res = m.Indexes.City.ExactMatch(ctx, "new york").PrimaryKeys()
-	require.Equal(t, []keys.Uint64Key{1, 2}, res)
+	require.Equal(t, []uint64{1, 2}, res)
 
 	// test ordinary map functionality
 	p, err := m.Get(ctx, 2)
@@ -68,6 +63,6 @@ func TestIndexedMap(t *testing.T) {
 	p = m.GetOr(ctx, 10, person{10, "sf"})
 	require.Equal(t, p, person{10, "sf"})
 
-	persons := m.Iterate(ctx, keys.NewRange[keys.Uint64Key]()).Values()
+	persons := m.Iterate(ctx, Range[uint64]{}).Values()
 	require.Equal(t, []person{{1, "new york"}, {2, "new york"}}, persons)
 }