+++ collections/item.go
@@ -1,25 +1,15 @@
 package collections
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // itemKey is a constant byte key which maps an Item object.
-var itemKey = []byte{0x0}
+var itemKey uint64 = 0
 
 // NewItem instantiates a new Item instance.
-func NewItem[V any, PV interface {
-	*V
-	Object
-}](cdc codec.BinaryCodec, sk sdk.StoreKey, prefix uint8) Item[V, PV] {
-	return Item[V, PV]{
-		prefix:   []byte{prefix},
-		sk:       sk,
-		cdc:      newStoreCodec(cdc),
-		typeName: typeName(PV(new(V))),
-	}
+func NewItem[V any](sk sdk.StoreKey, namespace Namespace, valueEncoder ValueEncoder[V]) Item[V] {
+	return (Item[V])(NewMap[uint64, V](sk, namespace, uint64Key{}, valueEncoder))
 }
 
 // Item represents a state object which will always have one instance
@@ -28,46 +18,16 @@ func NewItem[V any, PV interface {
 //   - config
 //   - parameters
 //   - a sequence
-type Item[V any, PV interface {
-	*V
-	Object
-}] struct {
-	_        V
-	prefix   []byte
-	sk       sdk.StoreKey
-	cdc      storeCodec
-	typeName string
-}
-
-func (i Item[V, PV]) getStore(ctx sdk.Context) sdk.KVStore {
-	return prefix.NewStore(ctx.KVStore(i.sk), i.prefix)
-}
+//
+// It builds on top of a Map with a constant key.
+type Item[V any] Map[uint64, V]
 
 // Get gets the item V or returns an error.
-func (i Item[V, PV]) Get(ctx sdk.Context) (V, error) {
-	s := i.getStore(ctx)
-	bytes := s.Get(itemKey)
-
-	var v V
-	if bytes == nil {
-		return v, notFoundError(i.typeName, "item")
-	}
-
-	i.cdc.unmarshal(bytes, PV(&v))
-	return v, nil
-}
+func (i Item[V]) Get(ctx sdk.Context) (V, error) { return (Map[uint64, V])(i).Get(ctx, itemKey) }
 
 // GetOr either returns the provided default
 // if it's not present in state, or the value found in state.
-func (i Item[V, PV]) GetOr(ctx sdk.Context, def V) V {
-	got, err := i.Get(ctx)
-	if err != nil {
-		return def
-	}
-	return got
-}
+func (i Item[V]) GetOr(ctx sdk.Context, def V) V { return (Map[uint64, V])(i).GetOr(ctx, itemKey, def) }
 
 // Set sets the item value to v.
-func (i Item[V, PV]) Set(ctx sdk.Context, v V) {
-	i.getStore(ctx).Set(itemKey, i.cdc.marshal(PV(&v)))
-}
+func (i Item[V]) Set(ctx sdk.Context, v V) { (Map[uint64, V])(i).Insert(ctx, itemKey, v) }