+++ collections/values.go
@@ -1,0 +1,34 @@
+package collections
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/gogo/protobuf/proto"
+)
+
+// ProtoValueEncoder returns a protobuf value encoder given the codec.BinaryCodec.
+// It's used to convert a specific protobuf object into bytes representation and convert
+// the protobuf object bytes representation into the concrete object.
+func ProtoValueEncoder[V any, PV interface {
+	*V
+	codec.ProtoMarshaler
+}](cdc codec.BinaryCodec) ValueEncoder[V] {
+	return protoValueEncoder[V, PV]{
+		cdc: cdc,
+	}
+}
+
+type protoValueEncoder[V any, PV interface {
+	*V
+	codec.ProtoMarshaler
+}] struct {
+	cdc codec.BinaryCodec
+}
+
+func (p protoValueEncoder[V, PV]) Name() string          { return proto.MessageName(PV(new(V))) }
+func (p protoValueEncoder[V, PV]) Encode(value V) []byte { return p.cdc.MustMarshal(PV(&value)) }
+func (p protoValueEncoder[V, PV]) Stringify(v V) string  { return PV(&v).String() }
+func (p protoValueEncoder[V, PV]) Decode(b []byte) V {
+	v := PV(new(V))
+	p.cdc.MustUnmarshal(b, v)
+	return *v
+}