+++ x/pricefeed/keeper/keeper.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	"github.com/NibiruChain/nibiru/collections"
-	"github.com/NibiruChain/nibiru/collections/keys"
 
 	"github.com/tendermint/tendermint/libs/log"
 
@@ -27,11 +26,10 @@ type (
 		memKey     storetypes.StoreKey
 		paramstore paramtypes.Subspace
 
-		RawPrices     collections.Map[keys.Pair[common.AssetPair, keys.StringKey], types.PostedPrice, *types.PostedPrice]
-		CurrentPrices collections.Map[common.AssetPair, types.CurrentPrice, *types.CurrentPrice]
+		RawPrices     collections.Map[collections.Pair[common.AssetPair, sdk.AccAddress], types.PostedPrice]
+		CurrentPrices collections.Map[common.AssetPair, types.CurrentPrice]
 		// PriceSnapshots maps types.PriceSnapshot to the common.AssetPair of the snapshot and the creation timestamp as keys.Uint64Key.
-		// TODO(mercilex): maybe it's worth to create a keys.Timestamp key?
-		PriceSnapshots collections.Map[keys.Pair[common.AssetPair, keys.Uint64Key], types.PriceSnapshot, *types.PriceSnapshot]
+		PriceSnapshots collections.Map[collections.Pair[common.AssetPair, time.Time], types.PriceSnapshot]
 	}
 )
 
@@ -52,9 +50,15 @@ func NewKeeper(
 		memKey:     memKey,
 		paramstore: ps,
 
-		CurrentPrices:  collections.NewMap[common.AssetPair, types.CurrentPrice](cdc, storeKey, 0),
-		RawPrices:      collections.NewMap[keys.Pair[common.AssetPair, keys.StringKey], types.PostedPrice](cdc, storeKey, 1),
-		PriceSnapshots: collections.NewMap[keys.Pair[common.AssetPair, keys.Uint64Key], types.PriceSnapshot](cdc, storeKey, 2),
+		CurrentPrices: collections.NewMap(storeKey, 0, common.AssetPairKeyEncoder, collections.ProtoValueEncoder[types.CurrentPrice](cdc)),
+		RawPrices: collections.NewMap(storeKey, 1,
+			collections.PairKeyEncoder[common.AssetPair, sdk.AccAddress](common.AssetPairKeyEncoder, collections.AccAddressKeyEncoder),
+			collections.ProtoValueEncoder[types.PostedPrice](cdc),
+		),
+		PriceSnapshots: collections.NewMap(
+			storeKey, 2,
+			collections.PairKeyEncoder[common.AssetPair, time.Time](common.AssetPairKeyEncoder, collections.TimeKeyEncoder),
+			collections.ProtoValueEncoder[types.PriceSnapshot](cdc)),
 	}
 }
 
@@ -118,7 +122,7 @@ func (k Keeper) PostRawPrice(
 
 	// Sets the raw price for a single oracle instead of an array of all oracle's raw prices
 	newPostedPrice := types.NewPostedPrice(pair, oracle, price, expiry)
-	k.RawPrices.Insert(ctx, keys.Join(pair, keys.String(oracle.String())), newPostedPrice)
+	k.RawPrices.Insert(ctx, collections.Join(pair, oracle), newPostedPrice)
 	return nil
 }
 
@@ -180,7 +184,7 @@ func (k Keeper) GatherRawPrices(ctx sdk.Context, token0 string, token1 string) e
 	k.CurrentPrices.Insert(ctx, assetPair, currentPrice)
 
 	// Update the TWAP prices
-	k.PriceSnapshots.Insert(ctx, keys.Join(assetPair, keys.Uint64(uint64(ctx.BlockTime().UnixMilli()))), types.PriceSnapshot{
+	k.PriceSnapshots.Insert(ctx, collections.Join(assetPair, ctx.BlockTime()), types.PriceSnapshot{
 		PairId:      pairID,
 		Price:       currentPrice.Price,
 		TimestampMs: ctx.BlockTime().UnixMilli(),
@@ -291,18 +295,13 @@ func (k Keeper) GetCurrentTWAP(ctx sdk.Context, token0 string, token1 string,
 		inverseIsActive = true
 	}
 
-	// traverse snapshots of the pair in reverse order from currentTime - TWAPLookBackWindow => now
-	pair := keys.PairPrefix[common.AssetPair, keys.Uint64Key](assetPair)                  // only current AssetPair
-	fromTime := ctx.BlockTime().Add(-1 * k.GetParams(ctx).TwapLookbackWindow).UnixMilli() // earliest timestamp aka current time - TWAPLookBackWindow
-	toTime := ctx.BlockTime().UnixMilli()                                                 // current time
-	start := keys.PairSuffix[common.AssetPair](keys.Uint64(uint64(fromTime)))             // this turns it into a suffix
-	end := keys.PairSuffix[common.AssetPair](keys.Uint64(uint64(toTime)))
-	rng := keys.NewRange[keys.Pair[common.AssetPair, keys.Uint64Key]]().
-		Prefix(pair).
-		Start(keys.Inclusive(start)).
-		End(keys.Exclusive(end)) // current block shouldn't have a snapshot until EndBlock is run
-
-	snapshots := k.PriceSnapshots.Iterate(ctx, rng).Values()
+	snapshots := k.PriceSnapshots.Iterate(
+		ctx,
+		collections.PairRange[common.AssetPair, time.Time]{}.
+			Prefix(assetPair).                                                           // only results of provided asset pair
+			StartInclusive(ctx.BlockTime().Add(-1*k.GetParams(ctx).TwapLookbackWindow)). // start from earliest timestamp which is current time - twap lockback window
+			EndExclusive(ctx.BlockTime()),                                               // end at current time - exclusive
+	).Values()
 	if len(snapshots) == 0 {
 		// if there are no snapshots, return -1 for the price
 		return sdk.OneDec().Neg(), types.ErrNoValidTWAP
@@ -343,7 +342,7 @@ func calcTwap(ctx sdk.Context, snapshots []types.PriceSnapshot) (sdk.Dec, error)
 
 // GetCurrentPrices returns all current price objects from the store
 func (k Keeper) GetCurrentPrices(ctx sdk.Context) types.CurrentPrices {
-	return k.CurrentPrices.Iterate(ctx, keys.NewRange[common.AssetPair]()).Values()
+	return k.CurrentPrices.Iterate(ctx, collections.Range[common.AssetPair]{}).Values()
 }
 
 // GetRawPrices fetches the set of all prices posted by oracles for an asset
@@ -355,10 +354,5 @@ func (k Keeper) GetRawPrices(ctx sdk.Context, pairStr string) types.PostedPrices
 			Use pair, %v, instead`, pair.Inverse().String(), pairStr))
 	}
 
-	prefix := keys.PairPrefix[common.AssetPair, keys.StringKey](pair)
-	return k.
-		RawPrices.
-		Iterate(ctx,
-			keys.NewRange[keys.Pair[common.AssetPair, keys.StringKey]]().Prefix(prefix),
-		).Values()
+	return k.RawPrices.Iterate(ctx, collections.PairRange[common.AssetPair, sdk.AccAddress]{}.Prefix(pair)).Values()
 }